Get rid of incompatible_types warning in gtk_embed example

+++ examples/gtk_embed/main.c
@@ -44,11 +44,11 @@ on_button_clicked (GtkButton *button, gpointer user_data)
 static gboolean
 on_stage_button_press (ClutterStage *stage, ClutterEvent *event, gpointer user_data)
 {
-  gint x = 0;
-  gint y = 0;
+  gfloat x = 0;
+  gfloat y = 0;
   clutter_event_get_coords (event, &x, &y);
 
-  g_print ("Stage clicked at (%d, %d)\n", x, y);
+  g_print ("Stage clicked at (%f, %f)\n", x, y);
 
   return TRUE; /* Stop further handling of this event. */
 }
